Change API license to Apache 2.0
While the *software* is licensed under AGPL, it doesn't make much sense
to apply the same license to the API or to a document defining the API.
"Licensing" an API is in itself a weird thing and I'm not sure this is
even needed. But the spec being under a permissive license rather than
AGPL has the advantage that derivatives (e.g. generated clients) are
more free in their choice of license (otherwise, if you generate a
client from a copylefted API spec, you might have to copyleft your
client too).